requirements.txt
-e .
##
## jsii must be 1.14.1 or better on Python 3.7
## https://github.com/aws/aws-cdk/issues/11255
##
# jsii>=1.14.1
# cattrs>=1.1.1

##
## CDK Modules
##
aws-cdk-lib>=2.4.0
constructs>=10.0.0,<11.0.0
